% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{TBM}
\alias{TBM}
\title{Tuberculosis Models}
\description{
TB model
}
\examples{
TB <- TBM$new()
TB$plot()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{TBM in TMB code}

\item{\code{current_pars}}{the parameters used for model simulation}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{output}}{return model's simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TBM-new}{\code{TBM$new()}}
\item \href{#method-TBM-update_pars}{\code{TBM$update_pars()}}
\item \href{#method-TBM-plot}{\code{TBM$plot()}}
\item \href{#method-TBM-clone}{\code{TBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TBM-new"></a>}}
\if{latex}{\out{\hypertarget{method-TBM-new}{}}}
\subsection{Method \code{new()}}{
Init the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TBM$new(pars = NULL, tmax = 150, dt = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{A named vectors of parameters, must matched with the existing parameters (see \code{pars_moldova} for details)}

\item{\code{tmax}}{Maximum number of years to simulate}

\item{\code{dt}}{Time steps, default to 1}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TBM-update_pars"></a>}}
\if{latex}{\out{\hypertarget{method-TBM-update_pars}{}}}
\subsection{Method \code{update_pars()}}{
update model's parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TBM$update_pars(pars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{a named vector of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TBM-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TBM-plot}{}}}
\subsection{Method \code{plot()}}{
plot model dynamic
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TBM$plot(states = c("S", "AS1", "LS1", "IS1"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{states}}{list of model states to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{ggplot} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TBM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TBM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
